Break up objectify stub
